pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppt-generator"
version = "0.1.0"
description = "Universal PPT generation engine"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=0.6.23",
    "Pillow>=10.0.0",
    "PyYAML>=6.0.1",
    "click>=8.1.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0"
]
